Homework1JDBC

+++ Homework1JDBC/src/main/java/otus/dao/AuthorDao.java
@@ -16,5 +16,5 @@ public interface AuthorDao {
 
     List<Author> getAll();
 
-    int deleteById(long id);
+    long deleteById(long id);
 }
